Improve support of large log file on 32-bit host (#308)
* Better support large log file when -D_FILE_OFFSET_BITS=64 on 32-bit host

* Use off_t to define log file size uniformly

+++ muduo/base/AsyncLogging.h
@@ -21,7 +21,7 @@ class AsyncLogging : boost::noncopyable
  public:
 
   AsyncLogging(const string& basename,
-               size_t rollSize,
+               off_t rollSize,
                int flushInterval = 3);
 
   ~AsyncLogging()
@@ -63,7 +63,7 @@ class AsyncLogging : boost::noncopyable
   const int flushInterval_;
   bool running_;
   string basename_;
-  size_t rollSize_;
+  off_t rollSize_;
   muduo::Thread thread_;
   muduo::CountDownLatch latch_;
   muduo::MutexLock mutex_;
